+++ lib/circleci/coverage_reporter/abstract_reporter.rb
@@ -5,6 +5,16 @@ module CircleCI
     # @abstract Subclass and override {#name}, {#create_build_result} and {#create_current_result}
     #   to implement a custom Reporter class.
     class AbstractReporter
+      # @param path [String] relative path from artifacts dir to coverage directory
+      def initialize(path = self.class::DEFAULT_PATH)
+        @path = path
+      end
+
+      # @return [Boolean] whether it is active
+      def active?
+        File.directory?(File.join(CoverageReporter.configuration.artifacts_dir, path))
+      end
+
       # @param base_build [Build, nil]
       # @param previous_build [Build, nil]
       # @return [Report]
@@ -23,6 +33,8 @@ module CircleCI
       end
 
       private
+
+      attr_reader :path
 
       # @param build [Build, nil]
       # @return [AbstractResult, nil]