+++ lib/circleci/coverage_reporter/runner.rb
@@ -5,12 +5,9 @@ module CircleCI
     class Runner
       # @return [void]
       def run
-        dump
         reports = reporters.map { |reporter| reporter.report(base_build, previous_build) }
         vcs_client.create_comment(reports)
       end
-
-      private
 
       # @return [void]
       def dump
@@ -22,6 +19,8 @@ base_build_number | #{base_build_number.inspect}
 previous_build    | #{previous_build.inspect}
         EOF
       end
+
+      private
 
       # @return [AbstractVCSClient]
       def vcs_client